+++ src/Utilities/Transpose.cpp
@@ -1,13 +1,14 @@
 #include <cstring>
+#include "../../includes/Utilities/Transpose.h"
 
-void transpose(float *A, unsigned N)
+void transpose(double *A, unsigned N)
 {
     unsigned i,j;
-    float B[N][N];
+    double B[N][N];
     for(i=0;i<N;i++)
         for(j=0;j<N;j++)
             B[j][i] =   A[i*N+j];
 
-    memcpy(A,&B[0][0],N*N*sizeof(float));
+    memcpy(A,&B[0][0],N*N*sizeof(double));
     return ;
 }
